fix: #2648, Validation plugin restructuring
