+++ besu-plugins/linea-sequencer/sequencer/src/main/java/net/consensys/linea/sequencer/txpoolvalidation/LineaTransactionPoolValidatorPlugin.java
@@ -15,6 +15,7 @@ import com.google.auto.service.AutoService;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.atomic.AtomicBoolean;
 import lombok.extern.slf4j.Slf4j;
 import net.consensys.linea.AbstractLineaRequiredPlugin;
 import net.consensys.linea.config.LineaRejectedTxReportingConfiguration;
@@ -22,6 +23,7 @@ import net.consensys.linea.config.LineaTransactionPoolValidatorCliOptions;
 import net.consensys.linea.jsonrpc.JsonRpcManager;
 import net.consensys.linea.plugins.config.LineaL1L2BridgeSharedConfiguration;
 import net.consensys.linea.sequencer.txpoolvalidation.metrics.TransactionPoolProfitabilityMetrics;
+import net.consensys.linea.sequencer.txvalidation.LineaBlockTransactionValidatorPlugin;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.plugin.BesuPlugin;
 import org.hyperledger.besu.plugin.ServiceManager;
@@ -39,6 +41,8 @@ import org.hyperledger.besu.plugin.services.transactionpool.TransactionPoolServi
 @Slf4j
 @AutoService(BesuPlugin.class)
 public class LineaTransactionPoolValidatorPlugin extends AbstractLineaRequiredPlugin {
+  public static final AtomicBoolean registered = new AtomicBoolean(false);
+
   private ServiceManager serviceManager;
   private TransactionPoolValidatorService transactionPoolValidatorService;
   private TransactionSimulationService transactionSimulationService;
@@ -48,6 +52,7 @@ public class LineaTransactionPoolValidatorPlugin extends AbstractLineaRequiredPl
 
   @Override
   public void doRegister(final ServiceManager serviceManager) {
+    registered.set(true);
     this.serviceManager = serviceManager;
 
     transactionPoolValidatorService =
@@ -70,6 +75,14 @@ public class LineaTransactionPoolValidatorPlugin extends AbstractLineaRequiredPl
 
   @Override
   public void doStart() {
+    if (LineaBlockTransactionValidatorPlugin.registered.get()) {
+      throw new IllegalStateException(
+          "Both LineaBlockTransactionValidatorPlugin and LineaTransactionPoolValidatorPlugin are"
+              + " enabled. Only one should be active at a time since their transaction type"
+              + " validation functionality overlaps. Use LineaTransactionPoolValidatorPlugin for"
+              + " RPC/P2P nodes or LineaBlockTransactionValidatorPlugin for validator nodes.");
+    }
+
     if (l1L2BridgeSharedConfiguration().equals(LineaL1L2BridgeSharedConfiguration.TEST_DEFAULT)) {
       throw new IllegalArgumentException("L1L2 bridge settings have not been defined.");
     }
@@ -95,6 +108,7 @@ public class LineaTransactionPoolValidatorPlugin extends AbstractLineaRequiredPl
                   worldStateService,
                   transactionSimulationService,
                   transactionPoolValidatorConfiguration(),
+                  transactionValidatorConfiguration(),
                   profitabilityConfiguration(),
                   tracerConfiguration(),
                   l1L2BridgeSharedConfiguration(),
@@ -171,6 +185,7 @@ public class LineaTransactionPoolValidatorPlugin extends AbstractLineaRequiredPl
   @Override
   public void stop() {
     super.stop();
+    registered.set(false);
     rejectedTxJsonRpcManager.ifPresent(JsonRpcManager::shutdown);
   }
 }
